+++ gui/src/app/pages/NwbPage/viewPlugins/Ephys/EphysSummaryItemView.tsx
@@ -20,6 +20,7 @@ import LazyPlotlyPlot from "../CEBRA/LazyPlotlyPlot";
 import { getJobOutputUrl, removeLeadingSlash } from "../CEBRA/PairioHelpers";
 import PairioItemView from "../CEBRA/PairioItemView";
 import ElectrodeGeometryView from "./ElectrodeGeometryView";
+import useTimeSeriesInfo from "../TimeSeries/useTimeseriesInfo";
 
 type Props = {
   width: number;
@@ -105,7 +106,7 @@ const EphysSummaryItemView: FunctionComponent<Props> = ({
 
   // const electricalSeriesPathChoices: string[] | undefined = useElectricalSeriesPathChoices(nwbFile);
 
-  const samplingRate = useElectricalSeriesSamplingRate(nwbFile, path);
+  const { samplingRate } = useTimeSeriesInfo(nwbFile, path);
 
   const tags = useMemo(() => ["neurosift", "EphysSummary"], []);
 
@@ -191,67 +192,6 @@ const EphysSummaryItemView: FunctionComponent<Props> = ({
       jobFilter={jobFilter}
     />
   );
-};
-
-export const useElectricalSeriesSamplingRate = (
-  nwbFile: RemoteH5FileX,
-  path: string,
-) => {
-  const [samplingRate, setSamplingRate] = useState<number | undefined | null>(
-    undefined,
-  );
-  useEffect(() => {
-    let canceled = false;
-    setSamplingRate(undefined);
-    (async () => {
-      try {
-        const es = await nwbFile.getGroup(path);
-        if (!es) throw Error(`Electrical series not found: ${path}`);
-        const dd = await nwbFile.getDataset(`${path}/data`);
-        if (!dd) throw Error(`Dataset not found: ${path}/data`);
-        const st = await nwbFile.getDataset(`${path}/starting_time`);
-        if (st) {
-          const rate = st.attrs["rate"];
-          if (canceled) return;
-          setSamplingRate(rate);
-        } else {
-          const ts = await nwbFile.getDataset(`${path}/timestamps`);
-          if (!ts)
-            throw Error(
-              `Dataset not found: ${path}/starting_time and ${path}/timestamps`,
-            );
-          const tsData = await nwbFile.getDatasetData(`${path}/timestamps`, {
-            slice: [[0, 1000]],
-          });
-          if (!tsData)
-            throw Error(`No data for timestamps: ${path}/timestamps`);
-          const rate = estimateSamplingRateFromTimestamps(
-            tsData as any as number[],
-          );
-          if (canceled) return;
-          setSamplingRate(rate);
-        }
-      } catch (err) {
-        console.error(err);
-        setSamplingRate(null);
-      }
-    })();
-    return () => {
-      canceled = true;
-    };
-  }, [nwbFile, path]);
-  return samplingRate;
-};
-
-const estimateSamplingRateFromTimestamps = (timestamps: number[]) => {
-  const diffs = timestamps
-    .filter((t) => !isNaN(t))
-    .slice(1)
-    .map((t, i) => t - timestamps[i]);
-  if (diffs.length === 0)
-    throw Error("No diffs when estimating sampling rate from timestamps");
-  const medianDiff = median(diffs);
-  return 1 / medianDiff;
 };
 
 // const useElectricalSeriesPathChoices = (nwbFile: RemoteH5FileX | null) => {